+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Fixed
+- CXV1-24132 - Combined updating and deleting user capacity rules into single function call with accompanying change to configurator
+
 ## [10.1.5] - 2021-07-08
 ## Fixed
 - CXV1-23171 - Start New Interaction duplicated 30 seconds apart. Added interaction id to outbound voice create request to prevent duplication
@@ -14,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [10.1.3] - 2021-06-23
 ## Fixed
+- CXV1-24132 - Combined updating and deleting user capacity rules into single function call
 - CXV1-24132 - Fix agent's capacity rule being set to null when capacity rule is being updated.
 
 ## [10.1.2] - 2021-06-15