+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [10.2.1] - 2021-07-26
+## Fixed
+- CXV1-24431 - Handle error for Salesforce set height initiation
+
 ## [10.2.0] - 2021-07-20
 ## Fixed
 - CXV1-24132 - Combined updating and deleting user capacity rules into single function call with accompanying change to configurator